requirements.txt
####### requirements.txt #######
boto3
onelogin
pyyaml
